+++ CHANGES.md
@@ -15,9 +15,7 @@
   - @melpon
 - [ADD] Linux で `v4l2` と `pipewire` の feature flag を同時に有効化可能にする
   - @melpon
-- [CHANGE] `VideoDevice`, `VideoDeviceList`, `VideoCapture` を構造体から trait に変更し、バックエンド別に具象型に分割する
-  - @melpon
-- [CHANGE] `VideoDeviceList::enumerate()` を `V4l2VideoDeviceList::enumerate()` 等のバックエンド別メソッドに変更する
+- [CHANGE] `VideoDevice`, `VideoDeviceList`, `VideoCapture` を構造体から enum に変更する
   - @melpon
 - [FIX] `VideoCapture`, `PixelBuffer` はスレッドセーフな構造体ではないので Sync を削除する
   - @melpon